+++ lib/maze/hooks/appium_hooks.rb
@@ -28,6 +28,10 @@ module Maze
         end
       end
 
+      def after_all
+        @client&.log_run_outro
+      end
+
       def at_exit
         @client&.stop_session
       end